+++ apps/party/CHANGELOG.md
@@ -1,5 +1,101 @@
 # @gtmgrid/party
 
+## 1.1.1
+
+### Patch Changes
+
+- @gtmgrid/auth@1.1.1
+
+## 1.1.0
+
+### Patch Changes
+
+- @gtmgrid/auth@1.1.0
+
+## 1.0.6
+
+### Patch Changes
+
+- @gtmgrid/auth@1.0.6
+
+## 1.0.5
+
+### Patch Changes
+
+- @gtmgrid/auth@1.0.5
+
+## 1.0.4
+
+### Patch Changes
+
+- @gtmgrid/auth@1.0.4
+
+## 1.0.3
+
+### Patch Changes
+
+- @gtmgrid/auth@1.0.3
+
+## 1.0.2
+
+### Patch Changes
+
+- @gtmgrid/auth@1.0.2
+
+## 1.0.1
+
+### Patch Changes
+
+- @gtmgrid/auth@1.0.1
+
+## 1.0.0
+
+### Patch Changes
+
+- @gtmgrid/auth@1.0.0
+
+## 0.22.12
+
+### Patch Changes
+
+- @gtmgrid/auth@0.22.12
+
+## 0.22.11
+
+### Patch Changes
+
+- @gtmgrid/auth@0.22.11
+
+## 0.22.10
+
+### Patch Changes
+
+- @gtmgrid/auth@0.22.10
+
+## 0.22.9
+
+### Patch Changes
+
+- @gtmgrid/auth@0.22.9
+
+## 0.22.8
+
+### Patch Changes
+
+- @gtmgrid/auth@0.22.8
+
+## 0.22.7
+
+### Patch Changes
+
+- @gtmgrid/auth@0.22.7
+
+## 0.22.6
+
+### Patch Changes
+
+- @gtmgrid/auth@0.22.6
+
 ## 0.22.5
 
 ### Patch Changes